Fixed overlap bug in edit.php
The textarea box was expanding all the way to the bottom of the page, but its z-index was lower than the buttons class so the bottom 75px of text was cut off. I wrapped the textarea in a div and added min- and max-height constraints so that the js textarea expander expands the textarea to the proper size.